backend/requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.2
numpy==2.3.5
packaging==25.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
scikit-learn==1.7.2
scipy==1.16.3
sqlparse==0.5.4
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
